+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### 🔄 Changed
 - Video and giphy attachments now use `makeMessageViewModifier`
 - Updated scalling of avatar images
+- Turn off channel updates when message thread shown
 
 ### 🐞 Fixed
 - `AttachmentTextView` respects configured body font